fix: implemented migration banner and refactored existing implementation
