+++ src/locales/en/common.json
@@ -54,6 +54,7 @@
   "An error occurred while sending your transaction to the network. Please try again.": "An error occurred while sending your transaction to the network. Please try again.",
   "An error occurred while signing your transaction. Please try again.": "An error occurred while signing your transaction. Please try again.",
   "An error occurred.": "An error occurred.",
+  "Announcement": "Announcement",
   "Anonymous analytics collection": "Anonymous analytics collection",
   "Appearance": "Appearance",
   "Application": "Application",
@@ -346,6 +347,7 @@
   "Lisk": "Lisk",
   "Lisk Desktop determines the country names using GeoLite2 data created by": "Lisk Desktop determines the country names using GeoLite2 data created by",
   "Lisk Terms of Use": "Lisk Terms of Use",
+  "Lisk V4 Migration": "Lisk V4 Migration",
   "Lisk Website": "Lisk Website",
   "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.": "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.",
   "Lisk {{version}}": "Lisk {{version}}",
@@ -465,6 +467,7 @@
   "Please enter your account password to backup the secret recovery phrase.": "Please enter your account password to backup the secret recovery phrase.",
   "Please enter your account password to restore this account.": "Please enter your account password to restore this account.",
   "Please enter your account password to sign this {{type}}.": "Please enter your account password to sign this {{type}}.",
+  "Please make sure to correctly migrate your nodes to the new network, in order to not miss any blocks after the network hard fork.": "Please make sure to correctly migrate your nodes to the new network, in order to not miss any blocks after the network hard fork.",
   "Please open the Lisk app on your hardware wallet device to see your accounts.": "Please open the Lisk app on your hardware wallet device to see your accounts.",
   "Please paste the JSON or drag and drop the file here.": "Please paste the JSON or drag and drop the file here.",
   "Please re-import your account.": "Please re-import your account.",
