+++ src/modules/account/components/AccountDetails/WelcomeView/index.js
@@ -8,6 +8,8 @@ import { PrimaryButton } from 'src/theme/buttons';
 import stylesPrimaryButton from '@theme/buttons/css/primaryButton.css';
 import routes from 'src/routes/routes';
 import stylesSecondary from '@theme/buttons/css/secondaryButton.css';
+import InfoBanner from '@common/components/infoBanner/infoBanner';
+import { INFO_BANNERS } from '@common/constants';
 import styles from './WelcomeView.css';
 
 function AccountUniqueValueProposition({ className, valueProposition, bulletPointNr }) {
@@ -44,58 +46,72 @@ export default function WelcomeView() {
   const { t } = useTranslation();
 
   return (
-    <section className={styles.welcomeView}>
-      <header className={styles.welcomeHeader}>
-        <div className={styles.logoContainer}>
-          <Icon name="liskLogoWhiteNormalized" height={32} />
-        </div>
-        <h1 className={styles.title}>{t('Welcome to Lisk')}</h1>
-        <p className={styles.description}>
-          {t(
-            'If you are new to the Lisk ecosystem, create an account by clicking on the “Create account”. If you have an account, then add it to your wallet by clicking on “Add account”.'
-          )}
-        </p>
-        <div className={styles.actionButtons}>
-          <DropdownButton
-            className={classNames(styles.dropdownWrapper, 'input-with-dropdown-dropdown')}
-            buttonClassName={classNames(
-              stylesPrimaryButton.button,
-              styles.createAccountBtn,
-              styles.inputDropdownButton
+    <div className={styles.wrapper}>
+      <InfoBanner
+        t={t}
+        show
+        name="walletPageBanner"
+        className={styles.bannerWrapper}
+        infoLabel={INFO_BANNERS.liskMigration.infoLabel(t)}
+        infoMessage={INFO_BANNERS.liskMigration.infoMessage(t)}
+        infoDescription={INFO_BANNERS.liskMigration.infoDescription(t)}
+        illustrationName={INFO_BANNERS.liskMigration.illustrationName}
+        infoLink={INFO_BANNERS.liskMigration.infoLink}
+        infoLinkText={INFO_BANNERS.liskMigration.infoLinkText}
+      />
+      <section className={styles.welcomeView}>
+        <header className={styles.welcomeHeader}>
+          <div className={styles.logoContainer}>
+            <Icon name="liskLogoWhiteNormalized" height={32} />
+          </div>
+          <h1 className={styles.title}>{t('Welcome to Lisk')}</h1>
+          <p className={styles.description}>
+            {t(
+              'If you are new to the Lisk ecosystem, create an account by clicking on the “Create account”. If you have an account, then add it to your wallet by clicking on “Add account”.'
             )}
-            buttonLabel={
-              <span>
-                {t('Create account')} <Icon name="arrowWhiteDown" />
-              </span>
-            }
-            size="s"
-            ButtonComponent={PrimaryButton}
-            align="right"
-            showArrow
-          >
-            <Link to={`${routes.register.path}?strength=${128}`}>12 word passphrase</Link>
-            <Link to={`${routes.register.path}?strength=${256}`}>24 word passphrase</Link>
-          </DropdownButton>
-          <Link
-            className={classNames(stylesSecondary.button, styles.addAccountBtn)}
-            to={routes.addAccountOptions.path}
-          >
-            {t('Add account')}
-          </Link>
-        </div>
-      </header>
-      <footer className={styles.welcomeFooter}>
-        <h2 className={styles.welcomeFooterTitle}>{t('Why do I need an account?')}</h2>
-        <div className={styles.accountUniqueValuePropositionListing}>
-          {accountUniqueValuePropositions.map((valueProposition, index) => (
-            <AccountUniqueValueProposition
-              key={index}
-              valueProposition={valueProposition}
-              bulletPointNr={index + 1}
-            />
-          ))}
-        </div>
-      </footer>
-    </section>
+          </p>
+          <div className={styles.actionButtons}>
+            <DropdownButton
+              className={classNames(styles.dropdownWrapper, 'input-with-dropdown-dropdown')}
+              buttonClassName={classNames(
+                stylesPrimaryButton.button,
+                styles.createAccountBtn,
+                styles.inputDropdownButton
+              )}
+              buttonLabel={
+                <span>
+                  {t('Create account')} <Icon name="arrowWhiteDown" />
+                </span>
+              }
+              size="s"
+              ButtonComponent={PrimaryButton}
+              align="right"
+              showArrow
+            >
+              <Link to={`${routes.register.path}?strength=${128}`}>12 word passphrase</Link>
+              <Link to={`${routes.register.path}?strength=${256}`}>24 word passphrase</Link>
+            </DropdownButton>
+            <Link
+              className={classNames(stylesSecondary.button, styles.addAccountBtn)}
+              to={routes.addAccountOptions.path}
+            >
+              {t('Add account')}
+            </Link>
+          </div>
+        </header>
+        <footer className={styles.welcomeFooter}>
+          <h2 className={styles.welcomeFooterTitle}>{t('Why do I need an account?')}</h2>
+          <div className={styles.accountUniqueValuePropositionListing}>
+            {accountUniqueValuePropositions.map((valueProposition, index) => (
+              <AccountUniqueValueProposition
+                key={index}
+                valueProposition={valueProposition}
+                bulletPointNr={index + 1}
+              />
+            ))}
+          </div>
+        </footer>
+      </section>
+    </div>
   );
 }
