+++ src/modules/common/components/infoBanner/infoBanner.css
@@ -44,6 +44,14 @@
   & .label {
     display: flex;
 
+    & > span.new {
+      background-color: var(--color-jade-green);
+    }
+
+    & > span.announcement {
+      background-color: var(--color-warning-border);
+    }
+
     & > span {
       background-color: var(--color-jade-green);
       color: var(--color-zodiac-blue);
@@ -175,6 +183,7 @@
   display: flex;
   align-items: inherit;
   text-decoration: underline;
+  color: var(--color-constant-white);
 
   & > img {
     margin-left: 8px;
@@ -189,6 +198,60 @@
 
     & .content {
       flex-basis: 100%;
+    }
+  }
+}
+
+.bannerSwiper {
+  background-color: var(--color-ultramarine-blue);
+  border-radius: var(--border-radius-standard);
+  margin-top: 32px;
+
+  & :global(.bannerWrapper) {
+    margin-top: 0;
+  }
+
+  & :global(.swiper-pagination-bullet) {
+    border: 1px solid var(--color-periwinkle-blue);
+
+    &:global(.swiper-pagination-bullet-active) {
+      background-color: white;
+    }
+  }
+
+  & :global {
+    & .swiper-button-next,
+    & .swiper-button-prev {
+      color: var(--color-constant-white);
+    }
+
+    & .swiper-button-next::after,
+    & .swiper-button-prev::after {
+      font-size: 32px;
+    }
+  }
+
+  & :global(.swiper-button-next) {
+    right: 5px;
+  }
+
+  & :global(.swiper-button-prev) {
+    left: 5px;
+  }
+
+  & :global(.swiper-pagination-bullets) {
+    text-align: left;
+    width: initial;
+    left: 30px;
+  }
+
+  & .bannerWrapper {
+    & p > a {
+      display: inline;
+      font-weight: var(--font-weight-bold);
+      font-size: 14px;
+      color: var(--color-constant-white);
+      cursor: pointer;
     }
   }
 }
